requirements.txt
asttokens==2.4.1
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==41.0.7
cssselect==1.2.0
decorator==5.1.1
dnspython==2.4.2
eventlet==0.33.3
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
Flask-MySQL==1.5.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
gunicorn==21.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
ipython==8.18.1
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mysql-connector-python==8.2.0
numpy==1.26.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
parsel==1.8.1
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.43
Protego==0.3.0
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.17.2
PyMySQL==1.1.0
pyOpenSSL==23.3.0
python-engineio==4.8.0
python-socketio==5.10.0
pyzbar==0.1.9
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.11.0
service-identity==23.1.0
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
tldextract==5.1.1
traitlets==5.14.0
Twisted==22.10.0
typing_extensions==4.9.0
urllib3==2.1.0
w3lib==2.1.2
wcwidth==0.2.12
Werkzeug==3.0.1
wsproto==1.2.0
zope.interface==6.1
